+++ LeetCode/LeetCodeDesign.h
@@ -10383,7 +10383,7 @@ public:
 
     bool deposit(int account, long long money) 
     {
-        if (account > m_account.size())
+        if (account > (int)m_account.size())
         {
             return false;
         }
